+++ catalog/app/model/graphql/schema.generated.ts
@@ -4792,6 +4792,15 @@ export default {
             },
             args: [],
           },
+          {
+            name: 'indexedContent',
+            type: {
+              kind: 'SCALAR',
+              name: 'String',
+              ofType: null,
+            },
+            args: [],
+          },
         ],
         interfaces: [],
       },